Start cleaning up helpers a bit.
